Fix definitions for cross compilers
