+++ pkg/config/default.go
@@ -7,7 +7,6 @@ import (
 	"path/filepath"
 	"regexp"
 	"strconv"
-	"syscall"
 
 	"github.com/containers/common/pkg/unshare"
 	"github.com/containers/storage"
@@ -112,22 +111,6 @@ const (
 	// SeccompDefaultPath defines the default seccomp path.
 	SeccompDefaultPath = _installPrefix + "/share/containers/seccomp.json"
 )
-
-const (
-	cgroupRoot         = "/sys/fs/cgroup"
-	_cgroup2SuperMagic = 0x63677270
-)
-
-// isCgroup2UnifiedMode returns whether we are running in cgroup 2 cgroup2 mode.
-func isCgroup2UnifiedMode() (isUnified bool, isUnifiedErr error) {
-	var st syscall.Statfs_t
-	if err := syscall.Statfs(cgroupRoot, &st); err != nil {
-		isUnified, isUnifiedErr = false, err
-	} else {
-		isUnified, isUnifiedErr = st.Type == _cgroup2SuperMagic, nil
-	}
-	return
-}
 
 // DefaultConfig defines the default values from containers.conf
 func DefaultConfig() (*Config, error) {
