+++ pkg/config/default_unsupported.go
@@ -1,0 +1,8 @@
+// +build !linux
+
+package config
+
+// isCgroup2UnifiedMode returns whether we are running in cgroup2 mode.
+func isCgroup2UnifiedMode() (isUnified bool, isUnifiedErr error) {
+	return false, nil
+}
